+++ pkg/interactive/posix.go
@@ -1,0 +1,9 @@
+// +build !windows
+
+package interactive
+
+import "os"
+
+func termHandle() (*os.File, error) {
+	return os.Open("/dev/tty")
+}