+++ pkg/output/stdout.go
@@ -27,28 +27,13 @@ const (
 
 type Stdoutput struct {
 	config  *ffuf.Config
-	Results []Result
-}
-
-type Result struct {
-	Input            map[string][]byte `json:"input"`
-	Position         int               `json:"position"`
-	StatusCode       int64             `json:"status"`
-	ContentLength    int64             `json:"length"`
-	ContentWords     int64             `json:"words"`
-	ContentLines     int64             `json:"lines"`
-	ContentType      string            `json:"content-type"`
-	RedirectLocation string            `json:"redirectlocation"`
-	Url              string            `json:"url"`
-	ResultFile       string            `json:"resultfile"`
-	Host             string            `json:"host"`
-	HTMLColor        string            `json:"-"`
+	Results []ffuf.Result
 }
 
 func NewStdoutput(conf *ffuf.Config) *Stdoutput {
 	var outp Stdoutput
 	outp.config = conf
-	outp.Results = []Result{}
+	outp.Results = []ffuf.Result{}
 	return &outp
 }
 
@@ -137,13 +122,28 @@ func (s *Stdoutput) Banner() {
 
 	// Print matchers
 	for _, f := range s.config.Matchers {
-		printOption([]byte("Matcher"), []byte(f.Repr()))
+		printOption([]byte("Matcher"), []byte(f.ReprVerbose()))
 	}
 	// Print filters
 	for _, f := range s.config.Filters {
-		printOption([]byte("Filter"), []byte(f.Repr()))
+		printOption([]byte("Filter"), []byte(f.ReprVerbose()))
 	}
 	fmt.Fprintf(os.Stderr, "%s\n\n", BANNER_SEP)
+}
+
+// Reset resets the result slice
+func (s *Stdoutput) Reset() {
+	s.Results = make([]ffuf.Result, 0)
+}
+
+// GetResults returns the result slice
+func (s *Stdoutput) GetResults() []ffuf.Result {
+	return s.Results
+}
+
+// SetResults sets the result slice
+func (s *Stdoutput) SetResults(results []ffuf.Result) {
+	s.Results = results
 }
 
 func (s *Stdoutput) Progress(status ffuf.Progress) {
@@ -175,9 +175,9 @@ func (s *Stdoutput) Info(infostring string) {
 		fmt.Fprintf(os.Stderr, "%s", infostring)
 	} else {
 		if !s.config.Colors {
-			fmt.Fprintf(os.Stderr, "%s[INFO] %s\n", TERMINAL_CLEAR_LINE, infostring)
+			fmt.Fprintf(os.Stderr, "%s[INFO] %s\n\n", TERMINAL_CLEAR_LINE, infostring)
 		} else {
-			fmt.Fprintf(os.Stderr, "%s[%sINFO%s] %s\n", TERMINAL_CLEAR_LINE, ANSI_BLUE, ANSI_CLEAR, infostring)
+			fmt.Fprintf(os.Stderr, "%s[%sINFO%s] %s\n\n", TERMINAL_CLEAR_LINE, ANSI_BLUE, ANSI_CLEAR, infostring)
 		}
 	}
 }
@@ -199,14 +199,18 @@ func (s *Stdoutput) Warning(warnstring string) {
 		fmt.Fprintf(os.Stderr, "%s", warnstring)
 	} else {
 		if !s.config.Colors {
-			fmt.Fprintf(os.Stderr, "%s[WARN] %s", TERMINAL_CLEAR_LINE, warnstring)
+			fmt.Fprintf(os.Stderr, "%s[WARN] %s\n", TERMINAL_CLEAR_LINE, warnstring)
 		} else {
 			fmt.Fprintf(os.Stderr, "%s[%sWARN%s] %s\n", TERMINAL_CLEAR_LINE, ANSI_RED, ANSI_CLEAR, warnstring)
 		}
 	}
 }
 
-func (s *Stdoutput) writeToAll(config *ffuf.Config, res []Result) error {
+func (s *Stdoutput) Raw(output string) {
+	fmt.Fprintf(os.Stderr, "%s%s", TERMINAL_CLEAR_LINE, output)
+}
+
+func (s *Stdoutput) writeToAll(filename string, config *ffuf.Config, res []ffuf.Result) error {
 	var err error
 	var BaseFilename string = s.config.OutputFile
 
@@ -218,37 +222,37 @@ func (s *Stdoutput) writeToAll(config *ffuf.Config, res []Result) error {
 	}
 
 	s.config.OutputFile = BaseFilename + ".json"
-	err = writeJSON(s.config, s.Results)
+	err = writeJSON(filename, s.config, s.Results)
 	if err != nil {
 		s.Error(err.Error())
 	}
 
 	s.config.OutputFile = BaseFilename + ".ejson"
-	err = writeEJSON(s.config, s.Results)
+	err = writeEJSON(filename, s.config, s.Results)
 	if err != nil {
 		s.Error(err.Error())
 	}
 
 	s.config.OutputFile = BaseFilename + ".html"
-	err = writeHTML(s.config, s.Results)
+	err = writeHTML(filename, s.config, s.Results)
 	if err != nil {
 		s.Error(err.Error())
 	}
 
 	s.config.OutputFile = BaseFilename + ".md"
-	err = writeMarkdown(s.config, s.Results)
+	err = writeMarkdown(filename, s.config, s.Results)
 	if err != nil {
 		s.Error(err.Error())
 	}
 
 	s.config.OutputFile = BaseFilename + ".csv"
-	err = writeCSV(s.config, s.Results, false)
+	err = writeCSV(filename, s.config, s.Results, false)
 	if err != nil {
 		s.Error(err.Error())
 	}
 
 	s.config.OutputFile = BaseFilename + ".ecsv"
-	err = writeCSV(s.config, s.Results, true)
+	err = writeCSV(filename, s.config, s.Results, true)
 	if err != nil {
 		s.Error(err.Error())
 	}
@@ -257,24 +261,33 @@ func (s *Stdoutput) writeToAll(config *ffuf.Config, res []Result) error {
 
 }
 
+// SaveFile saves the current results to a file of a given type
+func (s *Stdoutput) SaveFile(filename, format string) error {
+	var err error
+	switch format {
+	case "all":
+		err = s.writeToAll(filename, s.config, s.Results)
+	case "json":
+		err = writeJSON(filename, s.config, s.Results)
+	case "ejson":
+		err = writeEJSON(filename, s.config, s.Results)
+	case "html":
+		err = writeHTML(filename, s.config, s.Results)
+	case "md":
+		err = writeMarkdown(filename, s.config, s.Results)
+	case "csv":
+		err = writeCSV(filename, s.config, s.Results, false)
+	case "ecsv":
+		err = writeCSV(filename, s.config, s.Results, true)
+	}
+	return err
+}
+
+// Finalize gets run after all the ffuf jobs are completed
 func (s *Stdoutput) Finalize() error {
 	var err error
 	if s.config.OutputFile != "" {
-		if s.config.OutputFormat == "all" {
-			err = s.writeToAll(s.config, s.Results)
-		} else if s.config.OutputFormat == "json" {
-			err = writeJSON(s.config, s.Results)
-		} else if s.config.OutputFormat == "ejson" {
-			err = writeEJSON(s.config, s.Results)
-		} else if s.config.OutputFormat == "html" {
-			err = writeHTML(s.config, s.Results)
-		} else if s.config.OutputFormat == "md" {
-			err = writeMarkdown(s.config, s.Results)
-		} else if s.config.OutputFormat == "csv" {
-			err = writeCSV(s.config, s.Results, false)
-		} else if s.config.OutputFormat == "ecsv" {
-			err = writeCSV(s.config, s.Results, true)
-		}
+		err = s.SaveFile(s.config.OutputFile, s.config.OutputFormat)
 		if err != nil {
 			s.Error(err.Error())
 		}
@@ -288,30 +301,27 @@ func (s *Stdoutput) Result(resp ffuf.Response) {
 	if len(s.config.OutputDirectory) > 0 {
 		resp.ResultFile = s.writeResultToFile(resp)
 	}
-	// Output the result
-	s.printResult(resp)
-	// Check if we need the data later
-	if s.config.OutputFile != "" {
-		// No need to store results if we're not going to use them later
-		inputs := make(map[string][]byte, len(resp.Request.Input))
-		for k, v := range resp.Request.Input {
-			inputs[k] = v
-		}
-		sResult := Result{
-			Input:            inputs,
-			Position:         resp.Request.Position,
-			StatusCode:       resp.StatusCode,
-			ContentLength:    resp.ContentLength,
-			ContentWords:     resp.ContentWords,
-			ContentLines:     resp.ContentLines,
-			ContentType:      resp.ContentType,
-			RedirectLocation: resp.GetRedirectLocation(false),
-			Url:              resp.Request.Url,
-			ResultFile:       resp.ResultFile,
-			Host:             resp.Request.Host,
-		}
-		s.Results = append(s.Results, sResult)
+
+	inputs := make(map[string][]byte, len(resp.Request.Input))
+	for k, v := range resp.Request.Input {
+		inputs[k] = v
 	}
+	sResult := ffuf.Result{
+		Input:            inputs,
+		Position:         resp.Request.Position,
+		StatusCode:       resp.StatusCode,
+		ContentLength:    resp.ContentLength,
+		ContentWords:     resp.ContentWords,
+		ContentLines:     resp.ContentLines,
+		ContentType:      resp.ContentType,
+		RedirectLocation: resp.GetRedirectLocation(false),
+		Url:              resp.Request.Url,
+		ResultFile:       resp.ResultFile,
+		Host:             resp.Request.Host,
+	}
+	s.Results = append(s.Results, sResult)
+	// Output the result
+	s.PrintResult(sResult)
 }
 
 func (s *Stdoutput) writeResultToFile(resp ffuf.Response) string {
@@ -339,35 +349,35 @@ func (s *Stdoutput) writeResultToFile(resp ffuf.Response) string {
 	return fileName
 }
 
-func (s *Stdoutput) printResult(resp ffuf.Response) {
+func (s *Stdoutput) PrintResult(res ffuf.Result) {
 	if s.config.Quiet {
-		s.resultQuiet(resp)
+		s.resultQuiet(res)
 	} else {
-		if len(resp.Request.Input) > 1 || s.config.Verbose || len(s.config.OutputDirectory) > 0 {
+		if len(res.Input) > 1 || s.config.Verbose || len(s.config.OutputDirectory) > 0 {
 			// Print a multi-line result (when using multiple input keywords and wordlists)
-			s.resultMultiline(resp)
+			s.resultMultiline(res)
 		} else {
-			s.resultNormal(resp)
+			s.resultNormal(res)
 		}
 	}
 }
 
-func (s *Stdoutput) prepareInputsOneLine(resp ffuf.Response) string {
+func (s *Stdoutput) prepareInputsOneLine(res ffuf.Result) string {
 	inputs := ""
-	if len(resp.Request.Input) > 1 {
-		for k, v := range resp.Request.Input {
+	if len(res.Input) > 1 {
+		for k, v := range res.Input {
 			if inSlice(k, s.config.CommandKeywords) {
 				// If we're using external command for input, display the position instead of input
-				inputs = fmt.Sprintf("%s%s : %s ", inputs, k, strconv.Itoa(resp.Request.Position))
+				inputs = fmt.Sprintf("%s%s : %s ", inputs, k, strconv.Itoa(res.Position))
 			} else {
 				inputs = fmt.Sprintf("%s%s : %s ", inputs, k, v)
 			}
 		}
 	} else {
-		for k, v := range resp.Request.Input {
+		for k, v := range res.Input {
 			if inSlice(k, s.config.CommandKeywords) {
 				// If we're using external command for input, display the position instead of input
-				inputs = strconv.Itoa(resp.Request.Position)
+				inputs = strconv.Itoa(res.Position)
 			} else {
 				inputs = string(v)
 			}
@@ -376,30 +386,30 @@ func (s *Stdoutput) prepareInputsOneLine(resp ffuf.Response) string {
 	return inputs
 }
 
-func (s *Stdoutput) resultQuiet(resp ffuf.Response) {
-	fmt.Println(s.prepareInputsOneLine(resp))
+func (s *Stdoutput) resultQuiet(res ffuf.Result) {
+	fmt.Println(s.prepareInputsOneLine(res))
 }
 
-func (s *Stdoutput) resultMultiline(resp ffuf.Response) {
+func (s *Stdoutput) resultMultiline(res ffuf.Result) {
 	var res_hdr, res_str string
 	res_str = "%s%s    * %s: %s\n"
-	res_hdr = fmt.Sprintf("%s[Status: %d, Size: %d, Words: %d, Lines: %d]", TERMINAL_CLEAR_LINE, resp.StatusCode, resp.ContentLength, resp.ContentWords, resp.ContentLines)
-	res_hdr = s.colorize(res_hdr, resp.StatusCode)
+	res_hdr = fmt.Sprintf("%s[Status: %d, Size: %d, Words: %d, Lines: %d]", TERMINAL_CLEAR_LINE, res.StatusCode, res.ContentLength, res.ContentWords, res.ContentLines)
+	res_hdr = s.colorize(res_hdr, res.StatusCode)
 	reslines := ""
 	if s.config.Verbose {
-		reslines = fmt.Sprintf("%s%s| URL | %s\n", reslines, TERMINAL_CLEAR_LINE, resp.Request.Url)
-		redirectLocation := resp.GetRedirectLocation(false)
+		reslines = fmt.Sprintf("%s%s| URL | %s\n", reslines, TERMINAL_CLEAR_LINE, res.Url)
+		redirectLocation := res.RedirectLocation
 		if redirectLocation != "" {
 			reslines = fmt.Sprintf("%s%s| --> | %s\n", reslines, TERMINAL_CLEAR_LINE, redirectLocation)
 		}
 	}
-	if resp.ResultFile != "" {
-		reslines = fmt.Sprintf("%s%s| RES | %s\n", reslines, TERMINAL_CLEAR_LINE, resp.ResultFile)
+	if res.ResultFile != "" {
+		reslines = fmt.Sprintf("%s%s| RES | %s\n", reslines, TERMINAL_CLEAR_LINE, res.ResultFile)
 	}
-	for k, v := range resp.Request.Input {
+	for k, v := range res.Input {
 		if inSlice(k, s.config.CommandKeywords) {
 			// If we're using external command for input, display the position instead of input
-			reslines = fmt.Sprintf(res_str, reslines, TERMINAL_CLEAR_LINE, k, strconv.Itoa(resp.Request.Position))
+			reslines = fmt.Sprintf(res_str, reslines, TERMINAL_CLEAR_LINE, k, strconv.Itoa(res.Position))
 		} else {
 			// Wordlist input
 			reslines = fmt.Sprintf(res_str, reslines, TERMINAL_CLEAR_LINE, k, v)
@@ -408,9 +418,9 @@ func (s *Stdoutput) resultMultiline(resp ffuf.Response) {
 	fmt.Printf("%s\n%s\n", res_hdr, reslines)
 }
 
-func (s *Stdoutput) resultNormal(resp ffuf.Response) {
-	res := fmt.Sprintf("%s%-23s [Status: %s, Size: %d, Words: %d, Lines: %d]", TERMINAL_CLEAR_LINE, s.prepareInputsOneLine(resp), s.colorize(fmt.Sprintf("%d", resp.StatusCode), resp.StatusCode), resp.ContentLength, resp.ContentWords, resp.ContentLines)
-	fmt.Println(res)
+func (s *Stdoutput) resultNormal(res ffuf.Result) {
+	resnormal := fmt.Sprintf("%s%-23s [Status: %s, Size: %d, Words: %d, Lines: %d]", TERMINAL_CLEAR_LINE, s.prepareInputsOneLine(res), s.colorize(fmt.Sprintf("%d", res.StatusCode), res.StatusCode), res.ContentLength, res.ContentWords, res.ContentLines)
+	fmt.Println(resnormal)
 }
 
 func (s *Stdoutput) colorize(input string, status int64) string {